pyproject.toml
[project]
name = "google-news-trends-mcp"
version = "0.2.6"
description = "An MCP server to access RSS feeds of Google News and Google Trends"
readme = "README.md"
requires-python = ">=3.10.18"
authors = [{ name = "Jesse Manek", email = "jesse.manek@gmail.com" }]
keywords = ["google", "news", "rss", "trends", "mcp", "fastmcp", "llm", "nlp"]
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "asyncio>=3.4.3",
    "click>=8.2.1",
    "cloudscraper>=1.2.71",
    "fastmcp>=2.9.2",
    "gnews>=0.4.1",
    "googlenewsdecoder>=0.1.7",
    "lxml[html-clean]>=6.0.0",
    "fastapi>=0.111.0",
    "newspaper4k>=0.9.3.1",
    "nltk>=3.9.1",
    "playwright>=1.53.0",
    "pydantic>=2.11.7",
    "trendspy>=0.1.6",
    "uvicorn>=0.30.0",
    "PyJWT>=2.8.0",
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["auth", "main", "mcp_server", "tools"]

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.1",
]



[project.urls]
Homepage = "https://github.com/jmanek/google-news-trends-mcp"
Repository = "https://github.com/jmanek/google-news-trends-mcp"
Issues = "https://github.com/jmanek/google-news-trends-mcp/issues"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
